spring-study-18 Organize directory.
